Separate ArborX neighbor unit tests
