+++ core/unit_test/CMakeLists.txt
@@ -81,6 +81,9 @@ macro(Cabana_add_tests)
 endmacro()
 
 Cabana_add_tests(NAMES AoSoA Slice DeepCopy Tuple Sort LinkedCellList NeighborList Parallel)
+if(Cabana_ENABLE_ARBORX)
+  Cabana_add_tests(NAMES NeighborListArborX)
+endif()
 if(Cabana_ENABLE_MPI)
   Cabana_add_tests(MPI NAMES CommunicationPlan Distributor Halo)
 endif()
